adds task schemas

+++ datastores/postgres/local_docker/docker-entrypoint-initdb.d/init.sql/612_json_schemas.sql
@@ -1,6 +1,6 @@
 CREATE TABLE public.ai_json_schema_definitions(
     schema_id BIGINT NOT NULL DEFAULT next_id() PRIMARY KEY,
-    org_id int8 NOT NULL REFERENCES orgs(org_id),
+    org_id BIGINT NOT NULL REFERENCES orgs(org_id),
     is_obj_array bool NOT NULL DEFAULT false,
     schema_name text NOT NULL,
     schema_group text NOT NULL DEFAULT 'default'
@@ -11,7 +11,7 @@ CREATE INDEX ai_json_schema_definitions_org_ind ON public.ai_json_schema_definit
 ALTER TABLE "public"."ai_json_schema_definitions" ADD CONSTRAINT "ai_json_schema_definitions_name_org_uniq" UNIQUE ("org_id", "schema_name");
 
 CREATE TABLE public.ai_json_schema_fields(
-    schema_id int8 NOT NULL REFERENCES ai_json_schema_definitions(schema_id),
+    schema_id BIGINT NOT NULL REFERENCES ai_json_schema_definitions(schema_id),
     field_name text NOT NULL,
     field_description text NOT NULL,
     data_type text NOT NULL,
@@ -20,8 +20,8 @@ CREATE TABLE public.ai_json_schema_fields(
 CREATE INDEX idx_json_schema_field_id ON public.ai_json_schema_fields(schema_id);
 
 CREATE TABLE public.ai_json_task_schemas(
-    schema_id int8 NOT NULL REFERENCES ai_json_schema_definitions(schema_id),
-    task_id int8 NOT NULL REFERENCES ai_task_library(task_id),
+    schema_id BIGINT NOT NULL REFERENCES ai_json_schema_definitions(schema_id),
+    task_id BIGINT NOT NULL REFERENCES ai_task_library(task_id),
     PRIMARY KEY (schema_id, task_id)
 );
